resume_generator/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
openai>=1.50.0
httpx==0.25.1
pydantic==2.5.0
playwright==1.40.0
Pillow==10.1.0
jinja2==3.1.2
pypdf==3.17.0
python-multipart==0.0.6
